Day 22 Part 2 complete
Issue was that you can't assume that the shortest distance to a spot (regardless of tool) will be part of the optimal path - you might do better with a longer distance but using a different tool. So, store the tool as part of the distance table key rather than part of the value.